introduced (hopefully) solid solution for the management of children types during and after initialization of hard-coded and plugged type hierarchies; Types can be added as child either: * with a given relative priority * at a particular position number (for that moment of insertion!)
Still important: Take care of plugins with interfering markup within the same parent type!

Recommendation: Use the priorities also for hard-coded children orders, e.g:
type.addChildType(2, child1);
type.addChildType(3, child2);
type.addChildType(4, child3);

instead of;
type.addChildType(child1);
type.addChildType(child2);
type.addChildType(child3);

While both ways are working, the first method allows follow-up plugins to inject types to that parent more easily (e.g., type.add(2.5))

+++ KnowWE/KnowWE-core/src/main/java/de/knowwe/plugin/Plugins.java
@@ -126,7 +126,7 @@ public class Plugins {
 			if (!pathToRootContainsType(type.getPathToRoot(), pluggedType)) {
 				//
 				if (priority != null) {
-					type.addChildType(priority.intValue(), pluggedType);
+					type.addChildType(priority.doubleValue(), pluggedType);
 				}
 				else {
 					type.addChildType(0, pluggedType);
